+++ scrooge-generator-tests/src/test/resources/gold_file_output_scala/com/twitter/scrooge/test/gold/thriftscala/OverCapacityException.scala
@@ -72,7 +72,7 @@ object OverCapacityException extends ValidatingThriftStructCodec3[OverCapacityEx
   )
 
   override lazy val metaData: ThriftStructMetaData[OverCapacityException] =
-    new ThriftStructMetaData(this, structFields, fieldInfos, Nil, structAnnotations)
+    ThriftStructMetaData(this, structFields, fieldInfos, Nil, structAnnotations)
 
   /**
    * Checks that all required fields are non-null.