+++ scrooge-generator-tests/src/test/resources/gold_file_output_scala/com/twitter/scrooge/test/gold/thriftscala/PlatinumService.scala
@@ -362,7 +362,7 @@ object PlatinumService extends _root_.com.twitter.finagle.thrift.GeneratedThrift
       )
     
       override lazy val metaData: ThriftStructMetaData[Args] =
-        new ThriftStructMetaData(this, structFields, fieldInfos, Nil, structAnnotations)
+        ThriftStructMetaData(this, structFields, fieldInfos, Nil, structAnnotations)
     
       /**
        * Checks that all required fields are non-null.
@@ -644,7 +644,7 @@ object PlatinumService extends _root_.com.twitter.finagle.thrift.GeneratedThrift
       )
     
       override lazy val metaData: ThriftStructMetaData[Result] =
-        new ThriftStructMetaData(this, structFields, fieldInfos, Nil, structAnnotations)
+        ThriftStructMetaData(this, structFields, fieldInfos, Nil, structAnnotations)
     
       /**
        * Checks that all required fields are non-null.