+++ scrooge-generator-tests/src/test/resources/gold_file_output_scala/com/twitter/scrooge/test/gold/thriftscala/Recursive.scala
@@ -92,7 +92,7 @@ object Recursive extends ValidatingThriftStructCodec3[Recursive] with StructBuil
   )
 
   override lazy val metaData: ThriftStructMetaData[Recursive] =
-    new ThriftStructMetaData(this, structFields, fieldInfos, Nil, structAnnotations)
+    ThriftStructMetaData(this, structFields, fieldInfos, Nil, structAnnotations)
 
   /**
    * Checks that all required fields are non-null.