+++ scrooge-generator-tests/src/test/resources/gold_file_output_scala/com/twitter/scrooge/test/gold/thriftscala/ResponseUnion.scala
@@ -130,7 +130,7 @@ object ResponseUnion extends ValidatingThriftStructCodec3[ResponseUnion] {
     )
   )
 
-  override lazy val metaData = new ThriftStructMetaData(
+  override lazy val metaData = ThriftStructMetaData(
     this,
     Nil,
     Nil,